% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesky.R
\name{cholesky}
\alias{cholesky}
\title{Function to perform Cholesky factorization}
\usage{
cholesky(Q)
}
\arguments{
\item{Q}{A matrix}
}
\value{
The Cholesky decomposition, L, of Q
}
\description{
Function to perform Cholesky factorization
}
