% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesky_bandsparse.R
\name{cholesky_bandsparse}
\alias{cholesky_bandsparse}
\title{Function to perform Cholesky factorization on a band-sparse matrix}
\usage{
cholesky_bandsparse(Q, bandwidth)
}
\arguments{
\item{Q}{A matrix}

\item{bandwidth}{The bandwidth of Q}
}
\value{
The cholesky decomposition, L, of Q
}
\description{
Function to perform Cholesky factorization on a band-sparse matrix
}
