% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesky_bandsparse.R
\name{cholesky_bandsparse_Matrix}
\alias{cholesky_bandsparse_Matrix}
\title{Function to perform Cholesky factorization on a band-sparse matrix}
\usage{
cholesky_bandsparse_Matrix(Q)
}
\arguments{
\item{Q}{A matrix}
}
\value{
The cholesky decomposition, L, of Q
}
\description{
Function to perform Cholesky factorization on a band-sparse matrix
}
